ML/requirements.txt
Flask==3.0.3
tensorflow==2.10.1
tensorflow-addons==0.22.0
keras==2.10.0
numpy==1.26.4
h5py==3.12.1
protobuf==3.19.6
Werkzeug==3.0.6
pandas==1.5.3
gunicorn==20.1.0
Pillow==8.3.2
